auto test options and more
Added two autonomous options to test the gyroscope, got a (maybe) working deltatime function in auto to support one of the options, also, at the moment currentAutoTime is dependant on the accuracy of the deltatime thingy. That being said, I don't imagine that there'll be more than maybe a tenth of a second of drift total in a worst-case scenario. I also changed how the auto scoring option switches between bits of code (used to use selectedPosition, now there's just some more booleans and whatnot).